Added Interaction component
Added SInteractionComponent to project with one PrimaryInteract() function

PrimaryInteract() creates a line trace, detects if the trace hit object is of collision type WorldDynamic. Line trace is based off of character model's eye location and camera's rotation.

If the trace hits a WorldDynamic object that implements the SGameplayInterface, the Interact/PrimaryInteract() function can be executed.

Added an action mapping to the SCharacter class for PrimaryInteract.

Placed additional BP_ItemChest in level.